# docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: WhatsApp - WhatsApp API
  description: |
    <hr style="height: 5px;background: #007500;margin: 20px 0;box-shadow: 0px 3px 5px 0px rgb(204 204 204);"/>

    <div align="center">

    [![Telegram Group](https://img.shields.io/badge/Group-Telegram-%2333C1FF)](https://t.me/whatsappBR)
    [![Whatsapp Group](https://img.shields.io/badge/Group-WhatsApp-%2322BC18)](https://chat.whatsapp.com/HyO8X8K0bAo0bfaeW8bhY5)
    [![License](https://img.shields.io/badge/license-GPL--3.0-orange)](https://github.com/wandersonchaves/whatsapp-api/blob/main/LICENSE)
    [![Support](https://img.shields.io/badge/Buy%20me-coffe-orange)](https://app.picpay.com/user/wanderson.chaves)

    </div>
      
    <div align="center"><img src="https://raw.githubusercontent.com/wandersonchaves/whatsapp-api/main/public/images/cover.png"/></div>

    WhatsApp is an advanced WhatsApp integration API that seamlessly merges the popularity of WhatsApp with modern business applications. 
    With comprehensive features, the WhatsApp API empowers businesses to expand their digital presence, reach customers more directly, and 
    provide personalized and efficient support, all through the world's most widely used messaging platform.

    The WhatsApp API opens doors to more effective and engaging business communication, offering a modern approach to customer support, 
...
    can build stronger and more authentic relationships with their customers, generating lasting satisfaction and loyalty.

...
components:
  securitySchemes:
    apikeyAuth:
      type: apiKey
      in: header
      name: apikey
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
servers:
...
paths:
...
  /instance/create:
    post:
      tags:
        - Instance Controller
      summary: Create Instance
      description: |
        This endpoint is used to create a new instance.
        It receives a JSON object in the request body with the following fields:
        - instanceName: the name of the instance
        - description: an optional description of the instance
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                instanceName: whatsapp-bot
                description: 'Instance: Test V1'
      security:
        - apikeyAuth: []
      responses:
        '201':
          description: Created
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '442'
            ETag:
              schema:
                type: string
                example: W/"1ba-TGxN/leIZn7YtssW90IY6RKaHvM"
            Vary:
              schema:
                type: string
                example: Accept-Encoding
            Date:
              schema:
                type: string
                example: Mon, 04 Dec 2023 23:34:23 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                id: 1
                name: whatsapp
                description: 'Instance: Test V1'
                createdAt: '2023-12-04T00:00:00.000Z'
                updatedAt: '2023-12-04T00:00:00.000Z'
                Auth:
                  id: 1
                  token: >-
                    eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpbnN0YW5jZU5hbWUiOiJjb2RlY2hhdCIsImFwaU5hbWUiOiJ3aGF0c2FwcC1hcGkiLCJ0b2tlbklkIjoiYjEwZDYzZWYtZGQ4Ni00NzFmLTljZmMtOGYxMTAxYTJjNGI3IiwiaWF0IjoxNzAxNzMyODYyLCJleHAiOjE3MDE3MzY0NjIsInN1YiI6ImctdCJ9.ICtbfIIUm7tE52GlxhBTEHiYwWSNAeWTg3er6v1I9Qw
  /instance/fetchInstances:
    get:
      tags:
        - Instance Controller
      summary: Fetch Instances
      description: |
        This endpoint is used to fetch one or all instances. It accepts an optional query parameter `instanceName` to filter the results.
      security:
        - apikeyAuth: []
      parameters:
        - name: instanceName
          in: query
          schema:
            type: string
          description: Retrieve one or all instances - optional
          example: whatsapp-bot
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '631'
            ETag:
              schema:
                type: string
                example: W/"277-xtshIWa2qiaPH/lRSdSpOLf18eQ"
            Vary:
              schema:
                type: string
                example: Accept-Encoding
            Date:
              schema:
                type: string
                example: Tue, 05 Dec 2023 00:42:54 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              examples:
                example-0:
                  summary: 200 - Fetch Single
                  value:
                    id: 1
                    name: whatsapp
                    description: 'Instance: Test V1'
                    connectionStatus: ONLINE
                    ownerJid: null
                    profilePicUrl: null
                    createdAt: '2023-12-04T00:00:00.000Z'
                    updatedAt: '2023-12-04T00:00:00.000Z'
                    Auth:
                      id: 1
                      token: >-
                        eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpbnN0YW5jZU5hbWUiOiJjb2RlY2hhdCIsImFwaU5hbWUiOiJ3aGF0c2FwcC1hcGkiLCJ0b2tlbklkIjoiYjEwZDYzZWYtZGQ4Ni00NzFmLTljZmMtOGYxMTAxYTJjNGI3IiwiaWF0IjoxNzAxNzMyODYyLCJleHAiOjE3MDE3MzY0NjIsInN1YiI6ImctdCJ9.ICtbfIIUm7tE52GlxhBTEHiYwWSNAeWTg3er6v1I9Qw
                      createdAt: '2023-12-04T00:00:00.000Z'
                      updatedAt: '2023-12-04T00:00:00.000Z'
                    Webhook: null
                    Chatwoot: null
                    Typebot: null
                example-1:
                  summary: 200 - Fetch All
                  value:
                    - id: 1
                      name: whatsapp
                      description: 'Instance: Test V1'
                      connectionStatus: ONLINE
                      ownerJid: null
                      profilePicUrl: null
                      createdAt: '2023-12-04T00:00:00.000Z'
                      updatedAt: '2023-12-04T00:00:00.000Z'
                      Auth:
                        id: 1
                        token: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpbnN0YW5jZU5hbWUiOiJjb2RlY2hhdCIsImFwaU5hbWUiOiJ3aGF0c2FwcC1hcGkiLCJ0b2tlbklkIjoiYjEwZDYzZWYtZGQ4Ni00NzFmLTljZmMtOGYxMTAxYTJjNGI3IiwiaWF0IjoxNzAxNzMyODYyLCJleHAiOjE3MDE3MzY0NjIsInN1YiI6ImctdCJ9.ICtbfIIUm7tE52GlxhBTEHiYwWSNAeWTg3er6v1I9Qw
                        createdAt: '2023-12-04T00:00:00.000Z'
                        updatedAt: '2023-12-04T00:00:00.000Z'
                      Webhook: null
                      Chatwoot: null
                      Typebot: null
...
  /instance/refreshToken/:
    put:
      tags:
        - JWT
      summary: Refresh Token
      description: Endpoint to update the authentication token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                oldToken: <string> - required
      security:
        - noauthAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...